Move direct deposit-network check from (SUB)UTXO to (SUB)ENTITIES
